Fixes
Fixed jumping sound & improved code as requested on the pr
