--- Assets/Scripts/InteractWithObject.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9d655e58fc8384049943325a97946a18
